Email, Teacher Rec Validation, Date Change
Made it so that an email field had to have a “.” and an “@“ sign. Also
made sure that if your grade was below 6th, you could list just one
teacher rec. Finally changed the school year field to always have
current school year and next school year (pivot on May).